+++ src/main/java/cn/kimming/bookadmin/service/impl/PublisherServiceImpl.java
@@ -1,8 +1,10 @@
 package cn.kimming.bookadmin.service.impl;
 
+import cn.kimming.bookadmin.mapper.BookMapper;
 import cn.kimming.bookadmin.mapper.PublisherMapper;
 import cn.kimming.bookadmin.pojo.Publisher;
 import cn.kimming.bookadmin.service.IPublisherService;
+import cn.kimming.bookadmin.util.AdminException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,6 +16,8 @@ import java.util.List;
 public class PublisherServiceImpl implements IPublisherService {
     @Autowired
     private PublisherMapper publisherMapper;
+    @Autowired
+    private BookMapper bookMapper;
     @Override
     public List<Publisher> findAll() {
         return publisherMapper.findAll();
@@ -35,6 +39,10 @@ public class PublisherServiceImpl implements IPublisherService {
     }
     @Override
     public void deleteById(Long id) {
+        long count = bookMapper.selectCountByPublisherId(id);
+        if (count > 0) {
+            throw new AdminException("删除失败, 该出版社被图书信息使用中");
+        }
         publisherMapper.deleteByPrimaryKey(id);
     }
 }
